Fix YamlFile marker bug
