+++ .github/dependabot.yml
@@ -1,4 +1,3 @@
-# ~~ Generated by projen. To modify, edit .projenrc.js and run "npx projen".
 
 version: 2
 updates:
