+++ .gitpod.yml
@@ -1,4 +1,3 @@
-# ~~ Generated by projen. To modify, edit .projenrc.js and run "npx projen".
 
 tasks:
   - name: Setup
