+++ .mergify.yml
@@ -1,4 +1,3 @@
-# ~~ Generated by projen. To modify, edit .projenrc.js and run "npx projen".
 
 pull_request_rules:
   - name: Automatic merge on approval and successful build
